Add files via upload
The main file is flaskblog.py, the others are templates needed for the pages functionality.
The key thing needed to be added is a navbar, a login page ( which i hope to work on this week) and links between any other pages created, right now a homepage template exists and you can access it through http://localhost:5000/home and http://localhost:5000/register for the registration page. Im not sure how this will work on your machines with your MySQL database. I created a database called "myflaskapp"(the names of certain things need changed as i was following a guide to create this) with user "root" and password "test". i can create an account successfully and the error messages  work, the passwords and encrypted with Sha-256, although i am not sure how to check MySQL to see if the accounts created are stored there or not. But the database is there and ther is a connection between Flask and MySQL , i will discuss this in more detail in the meeting.